cambios

+++ src/main/java/com/diegoygabriela/backend_novalink/service/Inter/UsuarioService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface UsuarioService {
     void insert(Usuario usuario);
     List<Usuario> list();
-    void delete(Long idUsuario);
-    Usuario listId(Long idUsuario);
+    void delete(Integer idUsuario);
+    Usuario listId(Integer idUsuario);
     Usuario findByUsername(String username);
 }
